Changed RectangleObstacle.png to rectangleObstacle.png and fixed a huge problem
Apparently all the obstacles were having a clusterf*ck over on the side
of the screen. I screwed up on one of my for loops (in updateEntites) so
that it didn't run. Since I did for(var i = 0; i < liveObstacles; i++)
the problem being that I DIDN'T TELL IT TO ITERATE FOR THE LENGTH OF THE
ARRAY.

TODO:
- Get collisions to resolve
- Get obstacles to glow on contact
- Throw a party because I killed that bug.